Add files via upload
This is a supply chain solution built solely to solve the transparency problem in the supply chain. By implying this solution, the owner will have full transparency about the stage of the product, i.e., whether the product has been created( means deployed ), paid, or delivered.  

The Item. sol smart contract makes it easier for the buyer to make payment, where the smart contract generates a specific address for each new product and buyer has to transfer funds to that address. and the funds are automatically transferred to the ItemManager.sol smart contract, (i.e., the Owner).

"This is not an industry-ready solution, and neither am I a supply chain expert. I built this project based upon the knowledge I could extract from the internet."